Discovered that if the button is pressed and released quickly enough, the state change upon release will happen when the interrupt from the state change upon being pressed is being processed. This causes a BTN1 message to send but not followed by a BTN0 message despite the button being released.
Solution:	Added a 1 second timer after entering sleep mode while the button is pressed. This timer triggers an 	interrupt that records the state of the button and halts the timer. This also wakes the microcontroller from sleep. If a release state change was not recorded as described above, it will be processed after this interrupt is finished.